ci: Move the lint to its own stage and make it manual
Currently in sdk-ruby the check stage passes automatically,
because the lint stage is automatic and all of the functional tests
are set to manual/allow_failure, which does not block the pipeline
from passing.

First we move the lint check in a previous stage to isolate the tests.
However there's no way to tell gitlab "wait for *any* of the manual
jobs in this stage to pass".
You can only say "wait for all jobs" (allow_failure false on all jobs)
or "skip the stage entirely" (allow_failure true on all jobs).

We don't want to always run all the test variants,
so until Gitlab lets us say this we just make the lint job manual.
This forces the MR author to go trigger some jobs manually before
they can merge...